#11 - Organize model classes
