+++ AngularSonar/src/app/model/GeoSquare.ts
