+++ AngularSonar/src/app/model/HahSet.ts
