+++ AngularSonar/src/app/model/MapCoordinates.ts
