+++ AngularSonar/src/app/model/SonarClientData.ts
