+++ AngularSonar/src/app/model/SonarData.ts
