+++ AngularSonar/src/app/model/SonarState.ts
